requirements.txt
fastapi
uvicorn[standard]
python-multipart
easyocr
google-generativeai
python-dotenv
openpyxl
opencv-python
PyPDF2
pdf2image